+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [1.1.0] - 2026-01-15
+### Added
+- Font customization for figure title, legend title, and legend labels
+
+### Fixed
+- Remove field restrictions for `lib-prep-type` for compendia in the process config.
+
 ## [1.0.0] - 2025-12-16
 ### Added
 - Initial release